[scx_rusty] select_cpu to kick lower pri tasks
If we get to the end of select_cpu:
if no idle cpu, waking, and no other tasks are in the domain queue
then see if any cpu in the domain is running a task with lower
priority. Pick the one with the lowest priority, whose runtime
is also lower than the task we want to select the cpu for, and kick it.